CAMEL-21179 - Secret Properties Functions: Supporting secret name containing "/" - Hashicorp Vault
